Fixes CSS imports breaking on tests
Jest was not understanding CSS imports and the docs suggest using CSS
Module mocking to resolve the issue. I'm not entirely sure what it does
and how it will impact testing on more complex scenarios but the issue
seems to be resolved for now. As the use cases evolve I'll reevaluate
this decision to see if proxy mocks are the way to go. The docs mention
that it is the recommended approach for snapshot testing.

https://jestjs.io/docs/webpack#mocking-css-modules